+++ src/main/java/me/offsetmonkey538/baguette/BaguetteMain.java
@@ -6,8 +6,9 @@ import net.fabricmc.api.ModInitializer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class MainBaguette implements ModInitializer {
+public class BaguetteMain implements ModInitializer {
     public static final Logger LOGGER = LoggerFactory.getLogger("Baguette");
+    public static final String MOD_ID = "baguette";
 
     @Override
     public void onInitialize() {