+++ src/main/java/me/offsetmonkey538/baguette/config/BaguetteConfig.java
@@ -2,6 +2,7 @@ package me.offsetmonkey538.baguette.config;
 
 public record BaguetteConfig(
         int tntBaguetteExplosionStrength,
+        int chargedTntBaguetteExplosionStrength,
 
         int waterBaguetteHungerDurationTicks,
         int waterBaguetteHungerAmplifier,